avoid getting floating pointed

+++ Content.Server/_ES/WarpDrive/Components/ESWarpDriveGameRuleComponent.cs
@@ -108,10 +108,13 @@ public sealed partial class ESWarpDriveGameRuleComponent : Component
     /// <summary>
     ///     The percentage the warp drive was charged at the last time a screen packet was sent out.
     ///     Saved and checked to ensure we don't send too many screen updates, and instead only do it if the
-    ///     percentage changes by enough (0.05 atm)
+    ///     percentage changes by enough (5% atm)
     /// </summary>
+    /// <remarks>
+    ///     Int to avoid getting float precision fucked
+    /// </remarks>
     [DataField]
-    public float LastScreenUpdatedChargePercentage = 0.0f;
+    public int LastScreenUpdatedChargePercentage = 0;
 
     /// <summary>
     ///     where it all goes
